initial sidebar show-hide on small screens
